backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
pydantic==2.8.2
httpx==0.27.2
python-dotenv==1.0.1
pytest==8.3.2
requests==2.32.3
